Crate som_value

Modules

convert

To convert values to types, and vice versa.

interned

The representation for interned strings. Made to work with som-core/interner. Values need to be able to use it, and som-core depends on som-value, so we’d have a circular dependency if this wasn’t here. It is in the scope of the crate (“how we represent values in SOM”) but is annoying if this crate is ever to be standalone. It’s a bit annoying for it not to be there, but it can be kicked out.

value

Shared value representation logic (NaN boxing really)

value_ptr

Class for storing a value itself as a typed pointer.

Macros

delegate_to_base_value

Macro used to make AST-specific and BC-specific Value type “inherit” behavior from the base value type. Rust could avoid this by inferring that a BaseValue and a Value are the same. …but I’m not sure there’s a way for me to inform it. Maybe in a future version.