som_value

Module value

Shared value representation logic (NaN boxing really)

Structs

BaseValue

Constants

BASE_TAG

Base bit pattern needed to be set for all tags.

BIG_INTEGER_TAG

Tag bits for the BigInteger type.

BOOLEAN_TAG

Tag bits for the Boolean type.

CANON_NAN_BITS

Canonical NaN representation (minimum bitfield to represent NaN).

CELL_BASE_TAG

Base bit pattern needed to be set for all tags that are GC-managed.

CHAR_TAG

INTEGER_TAG

Tag bits for the Integer type.

IS_PTR_PATTERN

Bit pattern used to quickly check if a given 64-bit value houses a pointer-type value.

NIL_TAG

Tag bits for the Nil type.

STRING_TAG

Tag bits for the String type.

SYMBOL_TAG

Tag bits for the Symbol type.

TAG_EXTRACTION

Bit pattern used to quickly extract the tag bits from a 64-bit value.

TAG_SHIFT

The amount of bits to shift tags in the correct position within a 64-bit value.