som_value::value

Struct BaseValue

#[repr(C)]
pub struct BaseValue { /* private fields */ }

Implementations

impl BaseValue

pub const TRUE: BaseValue

The boolean true value.

pub const FALSE: BaseValue

The boolean false value.

pub const NIL: BaseValue

The nil value.

pub const INTEGER_ZERO: Self

The integer 0 value.

pub const INTEGER_ONE: Self

The integer 1 value.

pub const fn new(tag: u64, value: u64) -> Self

pub fn new_boolean(value: bool) -> Self

Returns a new boolean value.

pub fn is_ptr_type(self) -> bool

Returns whether this value is a pointer type value.

pub unsafe fn as_something<PTR>(self) -> Option<PTR>
where PTR: From<u64>,

Returns it at an arbitrary pointer. Used for debugging.

Safety

“Don’t”

pub fn as_u64(self) -> u64

Return the value as its internal representation: a u64 type.

pub fn tag(self) -> u64

Returns the tag bits of the value.

pub fn payload(self) -> u64

Returns the payload bits of the value.

pub fn extract_gc_cell<Ptr>(self) -> Ptr
where Ptr: From<u64>,

pub fn extract_pointer_bits(self) -> u64

pub fn new_integer(value: i32) -> Self

Returns a new integer value.

pub fn new_double(value: f64) -> Self

Returns a new double value.

pub fn new_symbol(value: Interned) -> Self

Returns a new symbol value.

pub fn new_char(value: char) -> Self

pub fn new_big_integer<BigIntPtr>(value: BigIntPtr) -> Self
where u64: From<BigIntPtr>, BigIntPtr: Deref<Target = BigInt> + From<u64>,

Returns a new big integer value.

pub fn new_string<StringPtr>(value: StringPtr) -> Self
where u64: From<StringPtr>, StringPtr: Deref<Target = String> + From<u64>,

Returns a new string value.

pub fn is_big_integer(self) -> bool

Returns whether this value is a big integer.

pub fn is_string(self) -> bool

Returns whether this value is a string.

pub fn is_nil(self) -> bool

Returns whether this value is `nil``.

pub fn is_integer(self) -> bool

Returns whether this value is an integer.

pub fn is_double(self) -> bool

Returns whether this value is a double.

pub fn is_boolean(self) -> bool

Returns whether this value is a boolean.

pub fn is_boolean_true(self) -> bool

Returns whether or not it’s a boolean corresponding to true. NB: does NOT check if the type actually is a boolean.

pub fn is_boolean_false(self) -> bool

Returns whether or not it’s a boolean corresponding to false. NB: does NOT check if the type actually is a boolean.

pub fn is_symbol(self) -> bool

Returns whether this value is a symbol.

pub fn is_char(self) -> bool

pub fn as_big_integer<BigIntPtr>(self) -> Option<BigIntPtr>
where u64: From<BigIntPtr>, BigIntPtr: From<u64>,

Returns this value as a big integer, if such is its type.

pub fn as_string<StringPtr>(self) -> Option<StringPtr>
where StringPtr: From<u64> + Deref<Target = String>,

Returns this value as a string, if such is its type.

pub fn as_integer(self) -> Option<i32>

Returns this value as an integer, if such is its type.

pub fn as_double(self) -> Option<f64>

Returns this value as a double, if such is its type.

pub fn as_boolean(self) -> Option<bool>

Returns this value as a boolean, if such is its type.

pub fn as_char(self) -> Option<char>

pub fn as_boolean_unchecked(self) -> bool

Returns this value as a boolean, but without checking whether or not it really is one.

pub fn as_symbol(self) -> Option<Interned>

Returns this value as a symbol, if such is its type.

pub fn is_ptr<T, PTR>(&self) -> bool
where T: HasPointerTag, PTR: Deref<Target = T> + From<u64> + Into<u64>,

pub fn as_ptr<T: HasPointerTag, PTR>(&self) -> Option<PTR>
where PTR: Deref<Target = T> + From<u64> + Into<u64>,

pub unsafe fn as_ptr_unchecked<T: HasPointerTag, PTR>(&self) -> PTR
where PTR: Deref<Target = T> + From<u64> + Into<u64>,

Safety

Only use when the type of the pointer was previously checked.

pub fn Boolean(value: bool) -> Self

pub fn Integer(value: i32) -> Self

pub fn Double(value: f64) -> Self

pub fn Symbol(value: Interned) -> Self

pub fn Char(value: char) -> Self

pub fn BigInteger<BigIntPtr>(value: BigIntPtr) -> Self
where u64: From<BigIntPtr>, BigIntPtr: Deref<Target = BigInt> + From<u64>,

pub fn String<Ptr>(value: Ptr) -> Self
where u64: From<Ptr>, Ptr: Deref<Target = String> + From<u64>,

pub unsafe fn as_mut_ptr(&self) -> *mut BaseValue

Returns a pointer to the underlying data, given a reference to a Value type. Not actually unsafe in itself, but considered as such because it’s VERY dangerous unless used correctly. Why does it exist? Because GC needs to store mutable references to values to modify them when moving memory around. Most values are stored as &Value, so this function is convenient.

Safety

The value used as a reference must be long-lived: if it is dropped at any point before invoking this function, we’ll get undefined behavior. In practice for our cases, this means any reference passed to this function must be A POINTER TO THE GC HEAP.